renamed histogram -> frequency; added copyright license; add mod test
